+++ bootstrap_modal_forms/forms.py
@@ -2,5 +2,9 @@ from django import forms
 from bootstrap_modal_forms.mixins import PopRequestMixin, CreateUpdateAjaxMixin
 
 
-class BSModalForm(PopRequestMixin, CreateUpdateAjaxMixin, forms.ModelForm):
+class BSModalForm(PopRequestMixin, forms.Form):
+    pass
+
+
+class BSModalModelForm(PopRequestMixin, CreateUpdateAjaxMixin, forms.ModelForm):
     pass