+++ examples/forms.py
@@ -2,30 +2,35 @@ from django import forms
 from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from django.contrib.auth.models import User
 
+from bootstrap_modal_forms.forms import BSModalModelForm, BSModalForm
 from bootstrap_modal_forms.mixins import PopRequestMixin, CreateUpdateAjaxMixin
-from bootstrap_modal_forms.forms import BSModalForm
-
 from .models import Book
 
 
-class BookForm(BSModalForm):
+class BookFilterForm(BSModalForm):
+    type = forms.ChoiceField(choices=Book.BOOK_TYPES)
+
+    class Meta:
+        fields = ["type", "clear"]
+
+
+class BookModelForm(BSModalModelForm):
     publication_date = forms.DateField(
-        error_messages={'invalid': 'Enter a valid date in YYYY-MM-DD format.'}
+        error_messages={"invalid": "Enter a valid date in YYYY-MM-DD format."}
     )
 
     class Meta:
         model = Book
-        exclude = ['timestamp']
+        exclude = ["timestamp"]
 
 
-class CustomUserCreationForm(PopRequestMixin, CreateUpdateAjaxMixin,
-                             UserCreationForm):
+class CustomUserCreationForm(PopRequestMixin, CreateUpdateAjaxMixin, UserCreationForm):
     class Meta:
         model = User
-        fields = ['username', 'password1', 'password2']
+        fields = ["username", "password1", "password2"]
 
 
 class CustomAuthenticationForm(AuthenticationForm):
     class Meta:
         model = User
-        fields = ['username', 'password']
+        fields = ["username", "password"]