+++ examples/views.py
@@ -1,13 +1,20 @@
+from django.contrib.messages.views import SuccessMessageMixin
 from django.urls import reverse_lazy
 from django.views import generic
 
-from bootstrap_modal_forms.generic import (BSModalLoginView,
-                                           BSModalCreateView,
-                                           BSModalUpdateView,
-                                           BSModalReadView,
-                                           BSModalDeleteView)
-
-from .forms import BookForm, CustomUserCreationForm, CustomAuthenticationForm
+from bootstrap_modal_forms.generic import (
+    BSModalLoginView,
+    BSModalCreateView,
+    BSModalUpdateView,
+    BSModalReadView,
+    BSModalDeleteView,
+    BSModalFormView,
+)
+from .forms import (
+    BookModelForm,
+    CustomUserCreationForm,
+    CustomAuthenticationForm,
+    BookFilterForm)
 from .models import Book
 
 
@@ -16,10 +23,36 @@ class Index(generic.ListView):
     context_object_name = 'books'
     template_name = 'index.html'
 
+    def get_queryset(self):
+        qs = super().get_queryset()
+        if 'type' in self.request.GET:
+            qs = qs.filter(book_type=int(self.request.GET['type']))
+        return qs
+
+
+class BookFilterView(BSModalFormView):
+    template_name = 'examples/filter_book.html'
+    form_class = BookFilterForm
+
+    def form_valid(self, form):
+        if "clear" in self.request.POST:
+            # the user has clicked on the 'Clear' button
+            self.filter = ''
+        else:
+            # the user has filtered the list of books
+            self.filter = '?type=' + form.cleaned_data["type"]
+
+        # call the base form_valid (that will call the get_success_url)
+        response = super().form_valid(form)
+        return response
+
+    def get_success_url(self):
+        return reverse_lazy('index') + self.filter
+
 
 class BookCreateView(BSModalCreateView):
     template_name = 'examples/create_book.html'
-    form_class = BookForm
+    form_class = BookModelForm
     success_message = 'Success: Book was created.'
     success_url = reverse_lazy('index')
 
@@ -27,7 +60,7 @@ class BookCreateView(BSModalCreateView):
 class BookUpdateView(BSModalUpdateView):
     model = Book
     template_name = 'examples/update_book.html'
-    form_class = BookForm
+    form_class = BookModelForm
     success_message = 'Success: Book was updated.'
     success_url = reverse_lazy('index')