+++ tests/base.py
@@ -1,7 +1,8 @@
-import time
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from selenium import webdriver
-from selenium.common.exceptions import WebDriverException
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support.wait import WebDriverWait
 
 MAX_WAIT = 10
 
@@ -10,64 +11,32 @@ class FunctionalTest(StaticLiveServerTestCase):
 
     # Basic setUp & tearDown
     def setUp(self):
+        self._initialise_firefox()
+
+    def _initialise_firefox(self):
         self.browser = webdriver.Firefox()
+
+    def _initialise_chrome(self):
+        chrome_options = webdriver.ChromeOptions()
+        # disable warning when opening Chrome without admin rights on extension
+        chrome_options.add_experimental_option('useAutomationExtension', False)
+        self.browser = webdriver.Chrome(chrome_options=chrome_options)
 
     def tearDown(self):
         self.browser.quit()
 
-    def wait_for_text_in_body(self, *args, not_in=None):
-        start_time = time.time()
-        # Infinite loop
-        while True:
-            try:
-                body = self.browser.find_element_by_tag_name('body')
-                body_text = body.text
-                # Check that text is in body
-                if not not_in:
-                    for arg in args:
-                        self.assertIn(arg, body_text)
-                # Check there is no text in body
-                else:
-                    for arg in args:
-                        self.assertNotIn(arg, body_text)
-                return
-            except (AssertionError, WebDriverException) as e:
-                # Return exception if more than 10s pass
-                if time.time() - start_time > MAX_WAIT:
-                    raise e
-                # Wait for 0.5s and retry
-                time.sleep(0.5)
-
-    def wait_for_modal(self, modalID):
-        start_time = time.time()
-        # Infinite loop
-        while True:
-            try:
-                modal = self.browser.find_element_by_id(modalID)
-                return modal
-            except (AssertionError, WebDriverException) as e:
-                # Return exception if more than 10s pass
-                if time.time() - start_time > MAX_WAIT:
-                    raise e
-                # Wait for 0.5s and retry
-                time.sleep(0.5)
+    def wait_for(self, class_name=None, element_id=None, tag=None, xpath=None):
+        return WebDriverWait(self.browser, 20).until(
+            expected_conditions.element_to_be_clickable
+            ((By.ID, element_id) if element_id else
+             (By.CLASS_NAME, class_name) if class_name else
+             (By.TAG_NAME, tag) if tag else
+             (By.XPATH, xpath))
+        )
 
     def wait_for_table_rows(self):
-        start_time = time.time()
-        # Infinite loop
-        while True:
-            try:
-                table = self.browser.find_element_by_tag_name('table')
-                tbody = table.find_element_by_tag_name('tbody')
-                # Slice removes tr in thead
-                trs = tbody.find_elements_by_tag_name('tr')
-                return trs
-            except (AssertionError, WebDriverException) as e:
-                # Return exception if more than 10s pass
-                if time.time() - start_time > MAX_WAIT:
-                    raise e
-                # Wait for 0.5s and retry
-                time.sleep(0.5)
+        tbody = self.wait_for(xpath="//table/tbody")
+        return tbody.find_elements_by_tag_name('tr')
 
     def check_table_row(self, table_row, cells_count, cells_values):
         cells = table_row.find_elements_by_tag_name('td')